backend/requirements.txt
fastapi>=0.104.0
ollama>=0.1.0
uvicorn>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
sqlalchemy>=2.0.0
opencv-python>=4.8.0
pillow>=10.0.0
PyYAML>=6.0.0
